* config.sub: No more OS-driven substitution of -pc with sed.
I'm not sure why this was originally added. It's certainly not needed
anymore because the OS will never be duplicated onto the send of the
`basic_machine`.  If the user passed `unknown` or no vendor, this will
already be filled in. If they passed something more specific, it's
customary to respect that.

Signed-off-by: Ben Elliston <bje@gnu.org>